feat: padronize controller responses

+++ backend/src/product/product.service.ts
@@ -1,17 +1,18 @@
-import { forwardRef, HttpException, HttpStatus, Inject, Injectable } from '@nestjs/common';
+import { forwardRef, HttpException, HttpStatus, Inject, Injectable, Logger } from '@nestjs/common';
 import { CreateProductDto } from './dto/create-product.dto';
 import { InjectModel } from '@nestjs/mongoose';
 import { Product } from './entities/product.entity';
 import { Model, Types } from 'mongoose';
 import { CategoryService } from '../category/category.service';
 import { UpdateProductDto } from './dto/update-product.dto';
-import { S3Service } from '../s3/s3.service';
+import { S3Service } from '../s3/s3.service'
 
 @Injectable()
 export class ProductService {
   constructor(
     @InjectModel(Product.name) private productModel: Model<Product>,
-    @Inject(forwardRef(() => CategoryService))private readonly categoryService: CategoryService,
+    @Inject(forwardRef(() => CategoryService)) private readonly categoryService: CategoryService,
+    private readonly logger = new Logger(ProductService.name),
     private readonly s3Service: S3Service
   ) { }
 
@@ -26,6 +27,7 @@ export class ProductService {
     const image = await this.s3Service.upload(file)
     const createdProduct = new this.productModel(createProductDto);
     createdProduct.imageUrl = image;
+    createdProduct.createdAt = new Date()
     const savedProduct = await createdProduct.save();
 
     await this.categoryService.addProductToCategories(
@@ -37,16 +39,16 @@ export class ProductService {
   }
 
   async findById(id: string) {
-    const product = await this.productModel.findById(id).exec()
+    const product = await this.productModel.findById(id).exec();
     if (!product) {
-      throw new HttpException("Product not found", HttpStatus.NOT_FOUND)
+      throw new HttpException("Product not found", HttpStatus.NOT_FOUND);
     }
-    return product;
+    return product
   }
 
   async findAll() {
-    const products = await this.productModel.find().sort({ updateAt: -1 }).exec();
-    return products
+    const products = await this.productModel.find().sort({ updatedAt: -1 }).exec();
+    return products;
   }
 
   async update(id: string, updateProductDto: UpdateProductDto, file: Express.Multer.File) {
@@ -62,7 +64,10 @@ export class ProductService {
   }
 
   async delete(id: string) {
-    const product = await this.findById(id); 
+    const product = await this.productModel.findById(id).exec()
+    if (!product) {
+      throw new HttpException("Product not found", HttpStatus.NOT_FOUND)
+    }
     if (product.categoryIds && product.categoryIds.length > 0) {
       await this.categoryService.removeProductFromCategories(id, product.categoryIds);
     }
@@ -72,7 +77,7 @@ export class ProductService {
   async findManyByIds(ids: string[]) {
     return this.productModel.find({ _id: { $in: ids } }).exec();
   }
-  
+
   async removeCategoryFromProducts(categoryId: string, productIds: string[]) {
     await this.productModel.updateMany(
       { _id: { $in: productIds } },
